requirements.txt
requests
PyYAML
openpyxl
jira
reportlab
